requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
sentry-sdk>=2.0.0
psycopg2-binary==2.9.10
alembic==1.15.2
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
stripe==12.0.0
apscheduler==3.11.0
pydantic==2.10.6
email-validator==2.2.0
requests==2.32.3
aiohttp==3.12.15
slowapi==0.1.9
cryptography>=42.0.0
bleach==6.1.0
redis==5.0.1
python-dateutil==2.8.2
bcrypt==4.0.1